### 2023-03-28 #5306 @editorialbot
[PRE REVIEW]: TauFactor 2: A GPU accelerated python tool for microstructural analysis
**Submitting author:** @stke9 (<a href="http://orcid.org/0000-0002-7263-6724">Steven Kench</a>)
**Repository:** https://github.com/tldr-group/taufactor
**Branch with paper.md** (empty if default branch): Joss-paper-revisions
**Version:** v1.0.0
