### 2023-03-31 6442563
added DOIs and corrected typo
