### 2023-04-02 #5306 @stke9
Great, thanks both. I updated the DOIs and UK typo, but I think the editorial bot is pointing to the main branch instead of Joss-paper-revisions branch. I'm not sure whether it's possible to change this, or if it's okay for zhubonan to just find the updated paper there?
