### 2023-04-19 #5358 @alexsquires
Hi @stke9 before raising any technical points as issues on the repo itself I just wanted to check an admin point.

 I'm certain there is a straightforward explanation for this, I just want to do due diligence!

With regards to the point:

> Contribution and authorship: Has the submitting author (@stke9) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

Could clarify the contributions of the paper authors as going off the repo alone it is not clear.
