### 2023-04-25 tldr-group/taufactor#72 @alexsquires
JOSS review, docstrings
With regards to the review point:

`Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?`

Generally speaking, the level of documentation is good, I just had a couple of points I wanted to clarify.

1.  Is there a reason why some functions are given docstrings, and some are not? e.g in `taufactor.metrics`, `triple_phase_boundary` has no docstring, but the others do.
2. As you are using sphinx+readthedocs to generate your documentation, would it be possible to generate module-level documentation on the readthedocs site? (Though this is more "nice-to-have")
