### 2023-04-25 tldr-group/taufactor#71 @Yugal0
> Hi developers,
>
> More of a request than a "revision" but with reference to the point on the reviewer's checklist:
>
> `Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).`
>
> could a fully working example be provided? I was able to get the code to run using the tiff file in the repo using the documentation, but it is not clearly signposted that this is a good example case in the docs, and additionally, I had to preprocess the data to get the code to run.
>
> It seems like all the pieces of a simple example that a new user could run are nearly place, but currently, it takes a bit of fiddling to get taufactor to run. I would have thought this may also clear up some of the confusion in #69.

Yes, it will be helpful to me, as well #69
