### 2023-05-04 tldr-group/taufactor#74 @alexsquires
Joss review, reproducibility of performance claims
With regards to a number of points on the reviewers checklist for the JOSS submission:

- Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
-  Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
- Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

Performance claims are made in the paper and the data is presented. It looks impressive, but unless I have missed it, there are no scripts or example data to attempt to reproduce. If these could be provided, I would happily tick these all off!
