### 2023-05-04 #5358 @alexsquires
I've been over the code, and my comments are included in:

- https://github.com/openjournals/joss-reviews/issues/5358#issuecomment-1515036625
- https://github.com/tldr-group/taufactor/issues/74#issue-1695627214
- https://github.com/tldr-group/taufactor/issues/72#issue-1682957145
- https://github.com/tldr-group/taufactor/issues/71#issue-1682942752

I'm very happy to recommend taufactor for publication if these can be addressed.
