### 2023-05-04 54960ee
fix mpsolver printing, added examples
