### 2023-05-04 tldr-group/taufactor#71 @stke9
Thanks for pointing this out! We have added an examples notebook that walks through the key functionality of the package :)
