+++ docs/taufactor.rst
@@ -1,0 +1,40 @@
+API Reference
+=================
+
+Core module
+-----------
+
+taufactor.taufactor module
+--------------------------
+
+.. automodule:: taufactor.taufactor
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+taufactor.metrics module
+------------------------
+
+.. automodule:: taufactor.metrics
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+taufactor.utils module
+----------------------
+
+.. automodule:: taufactor.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: taufactor
+   :members:
+   :undoc-members:
+   :show-inheritance:
