+++ docs/taufactor.rst
@@ -1,22 +1,26 @@
 API Reference
 =================
 
+
 Core module
------------
+-----------------
 
 taufactor.taufactor module
---------------------------
+~~~~~~~~~~~~~~~~~
+
 
 .. automodule:: taufactor.taufactor
    :members:
    :undoc-members:
    :show-inheritance:
 
+
 Submodules
-----------
+-----------------
 
 taufactor.metrics module
-------------------------
+~~~~~~~~~~~~~~~~~
+
 
 .. automodule:: taufactor.metrics
    :members:
@@ -24,7 +28,8 @@ taufactor.metrics module
    :show-inheritance:
 
 taufactor.utils module
-----------------------
+~~~~~~~~~~~~~~~~~
+
 
 .. automodule:: taufactor.utils
    :members:
@@ -32,7 +37,7 @@ taufactor.utils module
    :show-inheritance:
 
 Module contents
----------------
+------------------
 
 .. automodule:: taufactor
    :members:
