+++ docs/index.rst
@@ -2,7 +2,7 @@ Welcome to TauFactor's documentation!
 ======================================
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 3
    :caption: Contents:
 
    readme
