+++ .readthedocs.yml
@@ -11,6 +11,6 @@ sphinx:
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-    version: 3.8
+    version: 3.9
     install:
         - requirements: docs/requirements.txt
