### 2023-05-05 tldr-group/taufactor#74 @isaacsquires
Hi Alex,

Thanks for raising this. This is a tricky one, as the tools that we compare TauFactor to require different environments and installations. It is only [PoreSpy](https://porespy.org/) that we compare to that is a Python based tool. Having a single executable script to reproduce the results is challenging, as it will depend from machine to machine. Therefore, in the paper, we instead just reference the tools and report our environment and machine configuration.

For reproducibility, I can direct you to the installation guides for each of the tools:
- [TauFactor original](https://www.mathworks.com/matlabcentral/fileexchange/57956-taufactor)
- [PoreSpy](https://porespy.org/installation.html)
- [PuMa](https://puma-nasa.readthedocs.io/en/latest/readme.html#installation)

Our reported results were from a single machine with the same example dataset, which is provided in [the GitHub repo](https://github.com/tldr-group/taufactor/blob/main/TF-500x500x500-1x2x6-v3.tiff.zip).

I hope that clears it up!
