### 2023-05-05 tldr-group/taufactor#74 @alexsquires
Thank you for the context.

I will try running the example data in these codes, and will tick these off the checklist assuming I can (approximately) verify the performance claims
