### 2023-05-05 tldr-group/taufactor#72 @isaacsquires
Thank you for raising this. We have added dosctrings for the missing modules and created API reference for the modules in the docs. This should be resolved by #75
