### 2023-05-23 tldr-group/taufactor#74 @isaacsquires
Hi @alexsquires

Thanks for pointing this out. When rerunning the porespy analysis, I found some dependencies have been updated and it now runs much quicker (and so does TauFactor since the new Pytorch update). I have attached the scripts used to run the time comparisons for the most up to date porespy and taufactor:
[comparison_scripts.zip](https://github.com/tldr-group/taufactor/files/11542465/comparison_scripts.zip).

The new porespy times and updated taufactor times have been added to the graph in #75
