### 2023-06-28 tldr-group/taufactor#77 @ma-sadeghi
JOSS review: Python version >= 3.9 
NumPy has [dropped support](https://numpy.org/neps/nep-0029-deprecation_policy.html) for Python 3.8 on April 14, 2023. There are some places that unsupport Python versions have been used: `setup.py`, `tox.ini`, etc.
