### 2023-06-29 f115ca1
ma and lex issue fixes
