+++ tests/test_taufactor.py
@@ -9,7 +9,7 @@ import torch as pt
 from tests.utils import *
 import numpy as np
 import matplotlib.pyplot as plt
-
+import torch
 #  Testing the main solver
 
 
@@ -63,11 +63,33 @@ def test_solver_on_strip_of_ones():
     S.solve()
     assert np.around(S.tau, decimals=5) == 1
 
+def test_solver_on_slanted_strip_of_ones():
+    """Run solver on a slanted strip of ones"""
+    l = 20
+    img = np.zeros([l, l+1, l+1]).reshape(1, l, l+1, l+1)
+    # t = 10
+    # img[:, :, 0:t, 0:t] = 1
+    for i in range(l):
+        img[:, i, i:i+2, i:i+2] = 1
+    S = tau.Solver(img, device=pt.device('cpu'))
+    S.solve()
+    assert np.around(S.tau, decimals=5) == 4
 #  Testing the periodic solver
 
 
+def test_deadend():
+    """Test deadend pore"""
+    solid  = np.zeros((10,50,50))
+    solid[:8, 25,25] = 1
+    # solve for tau
+    S = tau.Solver(solid)
+    S.solve()
+    print(S.tau, S.converged)
+    assert np.around(S.D_eff, decimals=5) == 0
+    assert S.tau == np.inf
+
 def test_periodic_solver_on_uniform_block():
-    """Run solver on a block of ones."""
+    """Run periodic solver on a block of ones."""
     l = 20
     img = np.ones([l, l, l]).reshape(1, l, l, l)
     img[:, :, 0] = 0
@@ -77,7 +99,7 @@ def test_periodic_solver_on_uniform_block():
 
 
 def test_periodic_solver_on_empty_block():
-    """Run solver on a block of zeros."""
+    """Run periodic solver on a block of zeros."""
     l = 20
     img = np.zeros([l, l, l]).reshape(1, l, l, l)
     img[:, 0] = 1
@@ -87,7 +109,7 @@ def test_periodic_solver_on_empty_block():
 
 
 def test_periodic_solver_on_strip_of_ones():
-    """Run solver on a strip of ones, 1/4 volume of total"""
+    """Run periodic solver on a strip of ones, 1/4 volume of total"""
     l = 20
     img = np.zeros([l, l, l]).reshape(1, l, l, l)
     t = 10
@@ -157,7 +179,6 @@ def test_surface_area_on_empty_block():
 
     assert sa == 0
 
-
 def test_surface_area_on_checkerboard():
     """Run surface area on checkerboard block"""
     l = 50
@@ -195,6 +216,7 @@ def test_surface_area_on_periodic_2d():
 
 
 def test_surface_area_interfactial_3ph():
+    """Run surface area 3 phases """
     l = 3
     img = np.zeros([l, l, l])
     img[1] = 1
@@ -204,6 +226,7 @@ def test_surface_area_interfactial_3ph():
 
 
 def test_tpb_2d():
+    """Run tpb on 3x3"""
     l = 3
     img = np.zeros([l, l])
     img[0] = 1
@@ -213,6 +236,8 @@ def test_tpb_2d():
 
 
 def test_tpb_3d_corners():
+    """Run tpb on 2x2"""
+
     l = 2
     img = np.zeros([l, l, l])
     img[0, 0, 0] = 1
@@ -224,6 +249,8 @@ def test_tpb_3d_corners():
 
 
 def test_tpb_3d_corners():
+    """Run tpb on 2x2 corners"""
+
     l = 2
     img = np.zeros([l, l, l])
     img[0] = 1
@@ -233,6 +260,7 @@ def test_tpb_3d_corners():
 
 
 def test_multiphase_and_solver_agree():
[…]
