+++ tox.ini
@@ -1,8 +1,9 @@
 [tox]
-envlist = py35, py36, py37, py38, flake8
+envlist = py35, py36, py37, py38, py39, flake8
 
 [travis]
 python =
+    3.9: py39
     3.8: py38
     3.7: py37
     3.6: py36
