### 2023-06-29 tldr-group/taufactor#77 @stke9
Can you elaborate on why this is a problem - users can have compatible versions of python and numpy, we don't necessarily want to force them to use python 3.9 right? I have updated tox to include 3.9 option though!
