### 2023-06-29 tldr-group/taufactor#77 @ma-sadeghi
I agree that we shouldn't force people to not use unsupported Python versions. It's that unsupported Python/NumPy versions pose potential security risks/bugs that won't be patched. So, it's a good practice to encourage users to use the Python/NumPy versions that are still getting patched. Also, as more libraries drop support for Python 3.8, users may face compatibility issues.

As a side note, it just occured to me that since you've already released v1.0.0, dropping support for unsupported Python/NumPy versions wouldn't affect existing users (since it'll be on the dev branch). Once you make a new release on PyPI, it'll be enforced.

Regarding `setup.py`, could you add Python 3.10 as well? I'm asking because the default Python version on a freshly installed Ubuntu and also the Anaconda Python distribution is 3.10.
