+++ setup.py
@@ -36,6 +36,8 @@ setup(
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
+        'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
     ],
     description="TauFactor is an application for calculating tortuosity factors from tomographic data",
     install_requires=requirements,
