### 2023-07-24 tldr-group/taufactor#77 @isaacsquires
Hi, we have included 3.9 and 3.10 in our latest release :)
