### 2023-07-24 #5358 @isaacsquires
Hi all, we have just released with the reviewer additions: [v1.1.0](https://github.com/tldr-group/taufactor/releases/tag/v1.1.0)

Here is the DOI for the Zenodo archive: [10.5281/zenodo.8177306](https://doi.org/10.5281/zenodo.8177306)

That should be everything, please let me know if you need anything else.
