### 2023-08-03 #5358 @xuanxu
The errors come from the math equations code using displaymode with `\dfrac`. It looks like that function is not fully supported by our compilation tool. [This PR](https://github.com/tldr-group/taufactor/pull/86) should fix it, and running `recommend-accept` after is merged should work (or at least don't display those same errors).
