### 2023-08-09 #5358 @isaacsquires
The 'calaculated' typo has been fixed by the latest PR. The plural in materials is intentional, 'materials science' and 'materials characterisation' is our preferred syntax.
